requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==23.2.0
blinker==1.9.0
boto3==1.37.18
botocore==1.37.18
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
dill==0.3.9
docker==7.1.0
dotenv==0.9.9
fastapi==0.115.11
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.69.2
h11==0.14.0
hfapi==0.1b0
httplib2==0.22.0
huggingface-hub==0.29.3
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kagglehub==0.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mock==4.0.3
multidict==6.2.0
multiprocess==0.70.17
narwhals==1.31.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pathos==0.3.3
pillow==11.1.0
platformdirs==4.3.7
pox==0.3.5
ppft==1.7.6.9
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
sagemaker==2.242.0
sagemaker-core==1.0.25
schema==0.7.7
six==1.17.0
smdebug-rulesconfig==1.0.1
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
streamlit==1.43.2
tblib==3.0.0
tenacity==9.0.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
zipp==3.21.0
